backend/requirements.txt
flask==3.0.0
quart==0.19.4
quart-cors==0.7.0
aiohttp==3.9.3
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
PyPDF2==3.0.1
python-magic==0.4.27